requirements.txt
anyio==4.9.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.0
asn1crypto==1.5.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2023.11.17
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
crispy-bootstrap5==2025.4
cryptography==45.0.5
cssselect2==0.7.0
daphne==4.2.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.2.4
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-redis==6.0.0
django-versatileimagefield==3.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docx==0.2.4
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.4
fqdn==1.5.1
freetype-py==2.3.0
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
msgpack==1.1.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.3
notebook_shim==0.2.4
numpy==2.2.5
openpyxl==3.1.5
oscrypto==1.3.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.2.1
platformdirs==4.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.25.1
pycparser==2.22
Pygments==2.19.1
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pyOpenSSL==25.1.0
pypdf==3.17.4
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.1.0
python-json-logger==3.3.0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.1
pyzmq==26.4.0
qrcode==7.4.2
RapidFuzz==3.13.0
redis==6.2.0
referencing==0.36.2
reportlab==4.4.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlPyCairo==0.3.0
rpds-py==0.24.0
Send2Trash==1.8.3
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.5.1
syncedlyrics==1.0.1
terminado==0.18.1
tinycss2==1.2.1
tornado==6.4.2
traitlets==5.14.3
Twisted==25.5.0
txaio==25.6.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.2
uri-template==1.3.0
uritools==4.0.2
urllib3==2.1.0
virtualenv==20.26.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
xhtml2pdf==0.2.13
zope.interface==7.2
